refactor: separate auth and user
