+++ api/domain/user.go
@@ -26,21 +26,6 @@ type PaymentMethod struct {
 	CardOwner  *string `json:"card_owner" example:"Freddy Mercury"`
 }
 
-type PasswordChangeRequest struct {
-	ID        int64  `gorm:"primaryKey"`
-	UserID    int64  `gorm:"not null"`
-	TokenHash string `gorm:"not null"`
-	User      User   `gorm:"constraint:OnDelete:CASCADE;"`
-}
-
-func NewPasswordChangeRequest(tokenHash string, userID int64) *PasswordChangeRequest {
-	return &PasswordChangeRequest{
-		ID:        GenID(),
-		UserID:    userID,
-		TokenHash: tokenHash,
-	}
-}
-
 func NewUser(email string, passwordHash string) *User {
 	return &User{
 		ID:           GenID(),
