+++ api/dto/user.dto.go
@@ -78,18 +78,3 @@ func (b *UserUpdateBody) IntoMap() map[string]any {
 	}
 	return result
 }
-
-type LoginBody struct {
-	Email    string `json:"email" example:"johndoe@example.com"`
-	Password string `json:"password" example:"VerySecurePassword"`
-}
-
-type ChangePasswordBody struct {
-	RequestID   int64  `json:"request_id"`
-	Token       string `json:"token"`
-	NewPassword string `json:"new_password"`
-}
-
-type RequestPasswordChangeBody struct {
-	Email string `json:"email" example:"johndoe@example.com"`
-}
